// server/src/Types/System.Lib.d.ts
export type SystemPlatform = "Win" | "Lin" | "NotSupported" | "Test";
export type BashColorString =
	| "Default"
	| "Red"
	| "Black"
	| "Green"
	| "Yellow"
	| "Blue"
	| "Magenta"
	| "Cyan"
	| "Light Gray"
	| "Dark Gray"
	| "Light Red"
	| "Light Green"
	| "Light Yellow"
	| "Light Magenta"
	| "Light Cyan"
	| "White";